Move utils to cli
