+++ cli/utils/blob_utils.py
