+++ cli/utils/convert_to_jpeg.py
